requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
msgpack==1.1.0
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
PyMySQL==1.1.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.4.0
